requirements-dev.txt
-e .
bumpversion
wheel
tox
twine

# Tests
pytest>=4,<5
pytest-django==3.7.0
pytest-pythonpath==0.7.3
pytest-flake8==1.0.4
pytest-cov==2.8.1
mock==3.0.5

# Docs

Sphinx
#alabaster
sphinx_rtd_theme
recommonmark
m2r
